helios: make helios-extra a separate layer
The contents in watchdog-nowayout are not only for the watchdog, so it
seems pretty confusion. It also make no sense to keep a sublayer like
this. So move up all the contents in it to the parent directory and
make helios-extra a separate layer for all the specific things for
helios project.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>